args.out <-> outname problem
Hello, I tried your pipeline recently and it crushed with an error soooo close to the end

At the line 155 where the Counterexpression function is called, for a filename argument it takes args.out, which is an output path specified by the user. Throughout the pipeline the outname ( basename of args.out ) is used, and all of the temporary files are saved to the working directory using the outname as a prefix for tmp folder. 

Thus, when Counterexpression is called with args.out, it looks for temporary files following incorrect paths:
from base.py
def Countexpression(filename, allelement, genenumber, cellnumber, hdf5):
    gene_seen = allelement

    whitelist={}
    o = gzip.open('%s_scTEtmp/o4/%s.bed.gz'%(filename, filename), 'rt')

Temporarily changing args.out to outname should solve this problem